refactor: typescript optimized
