+++ .husky/pre-commit
@@ -2,3 +2,5 @@
 . "$(dirname "$0")/_/husky.sh"
 
 npm run lint && npm run format && npm run build
+# Stage modified files
+git add $(git diff --name-only)
